added test cases to UserControllerTest

+++ src/main/java/com/example/demo/model/persistence/Item.java
@@ -1,18 +1,15 @@
 package com.example.demo.model.persistence;
 
-import java.math.BigDecimal;
-
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.JoinColumn;
-import javax.persistence.ManyToOne;
 import javax.persistence.Table;
 
-import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
+
+import java.math.BigDecimal;
 
 @Entity
 @Table(name = "item")
@@ -29,7 +26,7 @@ public class Item {
 	
 	@Column(nullable = false)
 	@JsonProperty
-	private BigDecimal price;
+	private int price;
 	
 	@Column(nullable = false)
 	@JsonProperty
@@ -77,10 +74,10 @@ public class Item {
 	}
 
 	public BigDecimal getPrice() {
-		return price;
+		return BigDecimal.valueOf(price);
 	}
 
-	public void setPrice(BigDecimal price) {
+	public void setPrice(int price) {
 		this.price = price;
 	}
